Try fix #294 by using ActiveSupport.on_load
