+++ lib/action_controller/wechat_responder.rb
@@ -63,15 +63,4 @@ module ActionController
       end
     end
   end
-
-  if defined? Base
-    class << Base
-      include WechatResponder
-    end
-  end
-  if defined? API
-    class << API
-      include WechatResponder
-    end
-  end
 end
